// Frontend/audit-support/src/api/models/getUserDTO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * AuditSupport
 * OpenAPI spec version: v1
 */
import type { PersonInitials } from './personInitials';
import type { Roles } from './roles';

export interface GetUserDTO {
  id?: string;
  initials?: PersonInitials;
  /** @nullable */
  login?: string | null;
  /** @nullable */
  email?: string | null;
  role?: Roles;
  isActive?: boolean;
}